src/backend/requirements.txt
--index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://pypi.org/simple

fastapi==0.115.12
google-genai==1.12.1
langchain==0.3.24
langchain-experimental==0.3.4
langchain-google-genai==2.1.3
neo4j==5.28.1
nltk==3.9.1
# peft==0.14.0
pypdf2==3.0.1
pymupdf4llm==0.0.22
python-dotenv==1.1.0
spacy==3.8.5
# torch==2.5.1+cu124
python-multipart==0.0.20
# transformers==4.51.1
uvicorn==0.34.2


# spacy.cli.download("en_core_web_sm")
# python -c "import nltk; nltk.download('punkt'); nltk.download('punkt_tab');"